+++ lib/hash_ish/add_instance_methods.rb
@@ -1,29 +1,29 @@
 class HashIsh
   class AddInstanceMethods
-    def add(klass, kwargs)
+    def add(hash_ish, kwargs)
       kwargs.each do |key, value|
-        define_method(klass, key, value)
+        define_method(hash_ish, key, value)
       end
     end
 
-    def define_method(klass, name, value)
+    def define_method(hash_ish, name, value)
       return_value = method_value(value)
 
-      klass.send(:define_method, name) do
+      hash_ish.send(:define_singleton_method, name) do
         return_value
       end
     end
 
-    def hash?(object)
+    def is_a_hash?(object)
       object.is_a? Hash
     end
 
-    def hash_ish(value)
-      HashIsh.new(value)
+    def hash_ish(hash)
+      HashIsh.new(hash)
     end
 
     def method_value(value)
-      hash?(value) ? hash_ish(value) : value
+      is_a_hash?(value) ? hash_ish(value) : value
     end
   end
 end