+++ spec/hash_ish/add_instance_methods_spec.rb
@@ -1,34 +1,40 @@
 require 'spec_helper'
-require 'pry'
 
 describe HashIsh::AddInstanceMethods do
   let(:adder) { HashIsh::AddInstanceMethods.new }
-  let(:klass) { Class.new }
+  let(:hash_ish) { HashIsh.new }
 
   context '#add' do
-    it 'with empty hash' do
-      expect { adder.add(klass, {}) }.
-        to change { klass.new.methods.length }.
+    it 'to change by 0' do
+      expect { adder.add(hash_ish, {}) }.
+        to change { hash_ish.methods.length }.
         by(0)
     end
 
-    it 'with 1 pair' do
-      expect { adder.add(klass, { a: 1 }) }.
-        to change { klass.new.methods.length }.
+    it 'to change by 1' do
+      expect { adder.add(hash_ish, { a: 1 }) }.
+        to change { hash_ish.methods.length }.
         by(1)
     end
 
-    it 'with 2 pairs' do
-      expect { adder.add(klass, { a: 1, b: 2 }) }.
-        to change { klass.new.methods.length }.
+    it 'to change by 2' do
+      expect { adder.add(hash_ish, { a: 1, b: 2 }) }.
+        to change { hash_ish.methods.length }.
         by(2)
     end
   end
 
+  context '#is_a_hash?' do
+    it 'is true' do
+      expect(adder.is_a_hash?({})).
+        to eq(true)
+    end
+  end
+
   context '#hash_ish' do
-    it 'with 1 pair hash' do
-      expect(adder.hash_ish({a: 1})).
-        to respond_to(:a)
+    it 'to be a HashIsh' do
+      expect(adder.hash_ish({})).
+        to be_instance_of(HashIsh)
     end
   end
 end