+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.0 (2015-07-22)
+
+Features:
+  - Can now add default values
+  - HashIsh now returns instance of self
+
 ## 0.2.0 (2015-07-22)
 
 Features: